+++ packages/shared/forks/ReactFeatureFlags.native-fb.js
@@ -12,10 +12,8 @@ import invariant from 'shared/invariant';
 import typeof * as FeatureFlagsType from 'shared/ReactFeatureFlags';
 import typeof * as FeatureFlagsShimType from './ReactFeatureFlags.native-fb';
 
-// Re-export dynamic flags from the fbsource version.
-export const {
-  debugRenderPhaseSideEffects,
-} = require('../shims/ReactFeatureFlags');
+// Uncomment to re-export dynamic flags from the fbsource version.
+// export const {} = require('../shims/ReactFeatureFlags');
 
 // The rest of the flags are static for better dead code elimination.
 export const enableUserTimingAPI = __DEV__;
